# design-docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EmSim API
  version: 1.0.0-draft
  description: |
    REST/JSON API тренажёра диспетчера ДДС (RFC-001 §5).
    Аутентификация — cookie `emsim_session` (ADR-008). Все ответы несут `X-Request-ID`
    и `Cache-Control: no-store`. Ошибки — единый конверт `Error`.
    Push — Server-Sent Events (ADR-007): `/my/stream`, `/lessons/{lessonId}/stream`; формат событий — `sse-events.schema.json`.
servers:
  - url: /api/v1
...
security:
  - cookie: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Вход по логину и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string, maxLength: 64 }
                password: { type: string, maxLength: 256 }
                workstation_no: { type: integer, minimum: 1, description: 'Номер РМ. Обязателен для обучаемого (иначе 422, details.field=workstation_no); для admin/instructor игнорируется. Привязка РМ к IP-адресу не реализована в этом срезе (slice-planning.md §2).' }
      responses:
        '200':
          description: Сессия создана; cookie emsim_session установлена (ADR-008)
          content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { description: 'Учётная запись заблокирована после серии неверных паролей (ADR-038, code=account_locked). Блокировка снимается сама через LOGIN_LOCKOUT_DURATION или администратором.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422':
          description: workstation_no отсутствует, неизвестен или соответствующее РМ неактивно — details.field=workstation_no, details.reason=required|unknown|inactive
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Завершить сессию
      description: Возвращает 204 и сбрасывает cookie без сессии, с некорректной или уже недействительной сессией — logout идемпотентен и не раскрывает, была ли сессия действительной. Ошибка хранилища/аудита возвращает 500 и не сбрасывает cookie, чтобы клиент не получил ложное подтверждение отзыва сессии.
      responses:
        '204': { description: Сессия удалена (или отсутствовала) }
        '500': { $ref: '#/components/responses/Internal' }
  /me:
    get:
      tags: [auth]
      summary: Текущий пользователь, роль, РМ
      responses:
        '200':
          description: OK; emsim_session переустановлена с актуальным серверным expires_at для sliding-session
          content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/password:
    post:
      tags: [auth]
      summary: Сменить свой пароль (ADR-038)
      description: Нужен текущий пароль. Новый — не короче PASSWORD_MIN_LENGTH и не совпадает с текущим. Снимает признак credentials_change_required и завершает все остальные сеансы пользователя; текущий остаётся. Пока признак стоит, только этот маршрут, GET /me и выход доступны, остальное отвечает 403 password_change_required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, maxLength: 256 }
                new_password: { type: string, minLength: 8, maxLength: 256 }
      responses:
        '204': { description: Пароль сменён }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: 'Валидация: details.field=current_password reason=incorrect, или details.field=password reason=too_short|too_long|same_as_current', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ------------------------------------------------------------ admin
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: emsim_session
...
  responses:
...
    Unauthorized: { description: Не аутентифицирован, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    RateLimited: { description: Слишком много попыток, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Internal: { description: Внутренняя ошибка, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, validation_failed, unauthorized, forbidden, not_found, conflict, rate_limited, payload_too_large, unsupported_media_type, lesson_stopped, item_closed, stale_seq, transition_not_allowed, comment_required, call_required, call_in_progress, call_not_active, call_missed, internal_error, command_id_conflict, recording_conflict, recording_deadline_passed, stale_revision, stale_recommendation, stale_draft, has_blocking_issues, unsupported_for_editor, not_enough_scenarios, dictation_busy, dictation_unavailable, maintenance_mode, account_locked, password_change_required]
            message: { type: string }
            details: { type: object, additionalProperties: true }
        request_id: { type: string }
    Role: { type: string, enum: [admin, instructor, trainee] }
    Level: { type: string, enum: [easy, medium, hard] }
...
    Me:
      type: object
      required: [user, session_expires_at]
      properties:
        user: { $ref: '#/components/schemas/User' }
        workstation: { $ref: '#/components/schemas/Workstation', nullable: true, description: Отсутствует у admin/instructor и у обучаемого без активного РМ в сессии }
        session_expires_at: { type: string, format: date-time }
    User:
      type: object
      required: [id, login, full_name, role, level, active]
      properties:
        id: { type: string, format: uuid }
        login: { type: string }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        service_code: { type: string, nullable: true }
        level: { $ref: '#/components/schemas/Level', description: 'Текущий уровень обучаемого. Новый пользователь всегда easy; меняется только применением рекомендации преподавателем (срез 10), не через admin API.' }
        active: { type: boolean }
        locked_until: { type: [string, 'null'], format: date-time, description: 'ADR-038: вход заблокирован до этого момента; null, если блокировки нет' }
        credentials_change_required: { type: boolean, description: 'ADR-038: пароль задан администратором и должен быть сменён при первом входе (PASSWORD_FORCE_CHANGE). Имя намеренно без слова password: тело ответа никогда его не содержит.' }
...
    Workstation:
      type: object
      required: [number]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        number: { type: integer, minimum: 1 }
        label: { type: string }
        ip_address: { type: string, nullable: true }
        active: { type: boolean, default: true }
...